RunActionWithContext
Modified version of QueryRunWorkflowOnResults.  Passes a reference pointing to the target entity (of the workflow in which the activity is running) to the action being run on the results. The action being called must include an input parameter entityreference called ContextEntity, and entitytype can be any entity, so long as it matches match that of the target entity.  This allows the action to access the target entities values, making it an vehicle pass parameters.

I have not added any validation to confirm the called process is an action, or that it contains a matching parameter. At this point, this is just for discussion with Aiden.